master_node/requirements.txt
fastapi==0.78.0
uvicorn==0.17.0
cassandra-driver==3.25.0
requests==2.28.1
numpy==1.23.5
tensorflow-cpu==2.13.0 
kafka-python==2.0.2
wsproto==1.0.0  
aiokafka==0.7.2 